Moved Class | unecessary comment
